+++ sixtyfps_compiler/build.rs
@@ -58,7 +58,7 @@ fn process_style(path: &Path) -> std::io::Result<String> {
                     .path()
                     .extension()
                     .map(|ext| {
-                        ext == std::ffi::OsStr::new("60") || ext == std::ffi::OsStr::new("svg")
+                        ext == std::ffi::OsStr::new("60") || ext == std::ffi::OsStr::new("svg") || ext == std::ffi::OsStr::new("svgz")
                     })
                     .unwrap_or_default()
         })