+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
  - Enable support for compressed SVG (.svgz).
+ - Viewer: added possibility to set shell callback handler with `--on <callback> <handler>`
 
 ### Fixed